backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
geoalchemy2>=0.14.4
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
pydantic[email]>=2.5.3
python-multipart>=0.0.20
Pillow>=10.2.0
geopandas>=0.14.2
pandas>=2.2.0
xlrd>=2.0.1
fiona>=1.9.5
shapely>=2.0.2
rasterio>=1.3.9
pyshp>=3.0.3
requests>=2.31.0
alembic>=1.18.4
openai>=1.30.0,<2.0.0
xarray>=2026.4.0
netCDF4>=1.7.4
